refactor: Move `Worker` instantiation next to implementation
